refactor(audit-labellisation): extrait ChecklistActions de la vue checklist
Deplace la barre d'actions de l'en-tete (bouton de demande de premiere
etoile, bouton de demande d'audit, modale de demande de labellisation et
l'etat d'ouverture associe) dans un composant ChecklistActions dedie qui
lit le cycle et le referentiel via useChecklist. ChecklistView ne recoit
plus le cycle en prop ; CriteresLabellisationView ne le transmet plus.
Aucun changement de comportement.
